LPS-130662 Rename
